refactor: consolidate duplicated provenance-record shape guards (review #500)
isRecordShaped/describeNonArrayPayload/isValidProvenanceRecord/
describeInvalidRecord/REQUIRED_PROVENANCE_RECORD_FIELDS were duplicated
byte-for-byte between upgrade/apply.ts and adapters/provenance-io.ts,
along with the multi-line rationale explaining why. Round 3 of this
review had to fix the occupiedOwnershipBoundary check in both copies —
exactly the drift cost the duplication was meant to avoid paying twice.

These predicates are pure functions over `unknown` with no filesystem
access, so they don't need to live behind either side of the
upgrade-engine/adapters layering boundary. Moved to the neutral
packages/cli/src/provenance/validate.ts, imported by both callers. Each
caller still assembles its own full error message around the shared
fragments — the surrounding text differs per caller and is pinned by
that caller's own tests, so message wording is unchanged.

ReadProjectFileFn stays duplicated between upgrade/types.ts and
scaffold/types.ts: it is a one-line function-signature alias with no
logic to drift, and its only natural common home among the directories
in scope here (provenance/) would misplace a generic file-IO type under
a provenance-specific module purely to save one alias — worse coupling
for no drift-safety gain. index.ts already documents this and exports
only one public name for it.
